feat(task): added resources and assemblers for interfaces layer.

+++ src/main/java/com/construtech/buildsphere/platform/operationsManagement/domain/model/commands/CreateTeamCommand.java
@@ -1,4 +1,4 @@
 package com.construtech.buildsphere.platform.operationsManagement.domain.model.commands;
 
-public record CreateTeamCommand(String teamName, String description, int project) {
+public record CreateTeamCommand(String teamName, String description, Long project) {
 }
