+++ src/main/java/com/construtech/buildsphere/platform/operationsManagement/interfaces/resources/CreateWorkerResource.java
@@ -5,6 +5,6 @@ public record CreateWorkerResource(
         String role,
         int hoursWorked,
         Long teamId,
-        int projectId
+        Long projectId
 ) {
 }
